exercicio com resolucao
