+++ modulo4/intro-express/package.json
@@ -4,14 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "exercicio1": "tsc && node ./build/exercicio1.js",
-    "exercicio2": "tsc && node ./build/exercicio2.js",
-    "exercicio3": "tsc && node ./build/exercicio3,js",
-    "exercicio4": "tsc && node ./build/exercicio4,js",
-    "exercicio5": "tsc && node ./build/exercicio5,js",
-    "exercicio6": "tsc && node ./build/exercicio6,js",
-    "exercicio7": "tsc && node ./build/exercicio7,js",
-    "exercicio8": "tsc && node ./build/exercicio8,js"
+    "start": "tsc && node ./build/index.js"
   },
   "keywords": [],
   "author": "",
