--- modulo4/intro-express/src/exercicio2.ts
@@ -1,7 +1,0 @@
-type User = {
-    id: number,
-    name: string,
-    phone: string,
-    email: string,
-    website: string
-}
