--- modulo4/intro-express/src/exercicio3.ts
@@ -1,7 +1,0 @@
-const usuarios: User = {
-    id: 1,
-    name: "matheus",
-    phone: "9999-9999",
-    email: "matheus",
-    website: "meusite"
-}
